# Initial work to reduce memory in GetCacheKey